+++ tests/testthat/test-cross-validation.R
@@ -48,9 +48,8 @@ test_that("Basic cross validation works", {
 })
 
 test_that("Cross validation in parallel with globals", {
-  # https://github.com/pbs-assess/sdmTMB/issues/127
-  skip_on_ci()
   skip_on_cran()
+  # https://github.com/pbs-assess/sdmTMB/issues/127
   d <- pcod
   spde <- make_mesh(d, c("X", "Y"), cutoff = 15)
   set.seed(2)
@@ -66,7 +65,6 @@ test_that("Cross validation in parallel with globals", {
 })
 
 test_that("Leave future out cross validation works", {
-  skip_on_ci()
   skip_on_cran()
   x <- sdmTMB_cv(
     present ~ 1,
@@ -94,7 +92,6 @@ test_that("Leave future out cross validation works", {
 })
 
 test_that("Cross validation with offsets works", {
-  skip_on_ci()
   skip_on_cran()
   skip_if_not_installed("future")
   skip_if_not_installed("future.apply")
@@ -232,7 +229,6 @@ test_that("Cross validation with offsets works", {
 })
 
 test_that("Delta model cross validation works", {
-  skip_on_ci()
   skip_on_cran()
   set.seed(1)
   out_tw <- sdmTMB_cv(